secure routes formation

+++ api/app/Http/Controllers/FormationController.php
@@ -33,30 +33,28 @@ class FormationController extends Controller
             return response()->json($ApiResponse->getResponse(), 400);
         }
 
-        // Check user's permissions
-        $userRole = DB::select('SELECT role.name FROM cooperative_user_role INNER JOIN role ON role.id = role_id AND user_id=? AND cooperative_id=?', [$User->id, Input::get("cooperative_id")]);
-        if (isset($userRole[0])) {
-            if ($userRole[0]->name != 'enseignant') {
-                $ApiResponse->setErrorMessage("Permission denied, you must be an 'enseignant'.");
-                return response()->json($ApiResponse->getResponse(), 403);
+        // check cooperative
+        $cooperative_id_formation = CooperativeUser::select('cooperative_id')->where('user_id', $User->id)->get()->toArray();
+        $cooperative_id_column = array_column($cooperative_id_formation, 'cooperative_id');
+        if (($key = array_search(Input::get("cooperative_id"), $cooperative_id_column)) !== FALSE) {
+            try {
+                DB::beginTransaction();
+                // Add formation
+                $Formation = Formation::create([
+                    "name" => Input::get("name"),
+                    'estimated_duration' => Input::get("estimated_duration"),
+                    'level' => Input::get("level"),
+                    'cooperative_id' => Input::get("cooperative_id")
+                ]);
+                $ApiResponse->setMessage("Your formation was created.");
+                DB::commit();
+            } catch (\PDOException $e) {
+                DB::rollBack();
+                $ApiResponse->setErrorMessage($e->getMessage());
             }
         }
-
-        try {
-            DB::beginTransaction();
-            // Add formation
-            $Formation = Formation::create([
-                "name" => Input::get("name"),
-                'estimated_duration' => Input::get("estimated_duration"),
-                'level' => Input::get("level"),
-                'cooperative_id' => Input::get("cooperative_id")
-            ]);
-            $ApiResponse->setMessage("Your formation was created.");
-            DB::commit();
-        } catch (\PDOException $e) {
-            DB::rollBack();
-            $ApiResponse->setErrorMessage($e->getMessage());
-        }
+        else
+            $ApiResponse->setErrorMessage("You must be part of the cooperative to add this formation.");
 
         if ($ApiResponse->getError())
             return response()->json($ApiResponse->getResponse(), 400);
@@ -129,10 +127,11 @@ class FormationController extends Controller
         }
         
         $Formation = Formation::where("id", Input::get("formation_id"));
-        $cooperative_id_formation = CooperativeUser::where('user_id', $User->id);
+        $cooperative_id_formation = CooperativeUser::select('cooperative_id')->where('user_id', $User->id)->get()->toArray();
+        $cooperative_id_column = array_column($cooperative_id_formation, 'cooperative_id');
 
-        if ($Formation->first() && $cooperative_id_formation->first()) {
-            if ($cooperative_id_formation->first()->cooperative_id == $Formation->first()->cooperative_id) {
+        if ($Formation->first()) {
+            if (($key = array_search($Formation->first()->cooperative_id, $cooperative_id_column)) !== FALSE) {
                 $formation = $Formation->first()->toArray();
                 $formation['chapters'] = Chapter::select('id', 'name', 'type')->where('formation_id', Input::get('formation_id'))->orderBy('order', 'asc')->get()->toArray();
                 for ($i = 0; isset($formation['chapters'][$i]); $i++) {
@@ -164,32 +163,29 @@ class FormationController extends Controller
         $User = \Request::get("User");
         $validator = Validator::make($request->post(), [
             'formation_id' => 'bail|required|numeric',
-            'cooperative_id' => 'required|numeric'
         ]);
 
         if ($validator->fails()) {
             $ApiResponse->setErrorMessage($validator->messages()->first());
             return response()->json($ApiResponse->getResponse(), 400);
         }
-
-        // Check user's permissions
-        $userRole = DB::select('SELECT role.name FROM cooperative_user_role INNER JOIN role ON role.id = role_id AND user_id=? AND cooperative_id=?', [$User->id, Input::get("cooperative_id")]);
-        if (isset($userRole[0])) {
-            if ($userRole[0]->name != 'enseignant') {
-                $ApiResponse->setErrorMessage("You have no right on this formation.");
-                return response()->json($ApiResponse->getResponse(), 403);
-            }
-        }
-
+        
         // Delete formation
         $Formation = Formation::where('id', Input::get('formation_id'));
         if ($Formation->first()) {
-            try {
-                $Formation->delete();
-                $ApiResponse->setMessage("Successfuly removed this formation.");
-            } catch (Exception $ex) {
-                $ApiResponse->setErrorMessage("Failed to remove this formation. Please try again.");
+            // check cooperative
+            $cooperative_id_formation = CooperativeUser::select('cooperative_id')->where('user_id', $User->id)->get()->toArray();
+            $cooperative_id_column = array_column($cooperative_id_formation, 'cooperative_id');
+            if (($key = array_search($Formation->first()->cooperative_id, $cooperative_id_column)) !== FALSE) {
+                try {
+                    $Formation->delete();
+                    $ApiResponse->setMessage("Successfuly removed this formation.");
+                } catch (Exception $ex) {
+                    $ApiResponse->setErrorMessage("Failed to remove this formation. Please try again.");
+                }
             }
+            else
+                $ApiResponse->setErrorMessage("You must be part of the cooperative to see this formation.");;
         } else
             $ApiResponse->setErrorMessage("Formation not found.");
 
@@ -197,5 +193,22 @@ class FormationController extends Controller
             return response()->json($ApiResponse->getResponse(), 400);
         else
             return response()->json($ApiResponse->getResponse(), 200);
+    }
+
+    public function subscribe(Request $request)
+    {
+        $User = \Request::get("User");
+        $ApiResponse = new ApiResponse();
+        $rules = [
+            'formation_id' => "bail|required|numeric",
+            'cooperative_id' => 'required|numeric'
+        ];
+
+        $validator = Validator::make($request->post(), $rules);
+
+        if ($validator->fails()) {
+            $ApiResponse->setErrorMessage($validator->messages()->first());
+            return response()->json($ApiResponse->getResponse(), 400);
+        }
     }
 }
